remove check problems with detekt
